Fix various scheduler and load average bugs
This commit addresses five specific logic bugs identified in the kernel scheduler:
1. PackageEntry::CoreWakesUp: Corrected the condition to detect when a package transitions to fully active (last core waking up).
2. CoreEntry::RemoveCPU: Fixed unbalanced fIdleCPUCount by decrementing it unconditionally upon CPU removal.
3. RunQueue::ConstIterator: Prevented 32-bit overflow in masking logic by using 64-bit constants.
4. scheduler_load: Fixed the load average formula by removing a redundant scaling factor.
5. ThreadData::ComputeQuantum: Ensured interactivity scaling does not drop the quantum below the responsive floor (floorQuantum).

+++ src/system/kernel/scheduler/RunQueue.h
@@ -250,7 +250,7 @@ RUN_QUEUE_CLASS_NAME::ConstIterator::_FindNextPriority()
 
 	if (fPriority > 0 && currentBit != 31) {
 		// Mask bits at currentBit+1 and above, keep bits 0..currentBit
-		val &= (1UL << (currentBit + 1)) - 1;
+		val &= (uint32)((1ULL << (currentBit + 1)) - 1);
 	} else if (fPriority == 0) {
 		// If we finished bit 0, this word is done.
 		val = 0;
